gcs-to-bq-trigger/requirements.txt
flask
google-cloud-bigquery

